+++ tests/Feature/Framework/FacadeAware/Module/UserInput/Command/TestHiCommand.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace GacelaTest\Feature\Framework\FacadeAware\Module\UserInput\Command;
 
-use Gacela\Framework\DocBlockResolverAwareTrait;
+use Gacela\Framework\ServiceResolverAwareTrait;
 use GacelaTest\Feature\Framework\FacadeAware\Module\Facade;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -15,7 +15,7 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 final class TestHiCommand extends Command
 {
-    use DocBlockResolverAwareTrait;
+    use ServiceResolverAwareTrait;
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {