+++ tests/Feature/Framework/UsingCustomSuffixTypes/ModuleB/CommandModuleB.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace GacelaTest\Feature\Framework\UsingCustomSuffixTypes\ModuleB;
 
-use Gacela\Framework\DocBlockResolverAwareTrait;
+use Gacela\Framework\ServiceResolverAwareTrait;
 
 /**
  * @method FacadeModuleB getFacade()
  */
 final class CommandModuleB
 {
-    use DocBlockResolverAwareTrait;
+    use ServiceResolverAwareTrait;
 
     public function execute(): array
     {