+++ tests/Feature/Framework/UsingCustomSuffixTypes/ModuleC/CommandModuleC.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace GacelaTest\Feature\Framework\UsingCustomSuffixTypes\ModuleC;
 
-use Gacela\Framework\DocBlockResolverAwareTrait;
+use Gacela\Framework\ServiceResolverAwareTrait;
 
 /**
  * @method Facade getFacade()
  */
 final class CommandModuleC
 {
-    use DocBlockResolverAwareTrait;
+    use ServiceResolverAwareTrait;
 
     public function execute(): array
     {