+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Added PHP attribute support for `DocBlockResolverAwareTrait`
+- Added `ServiceResolverAwareTrait` with caching improvements; will replace `DocBlockResolverAwareTrait`
+- Introduced the `#[ServiceMap]` attribute as the preferred service binding instead of `DocBlock`
 
 ## [1.11.0](https://github.com/gacela-project/gacela/compare/1.10.0...1.11.0) - 2025-10-12